requirements.txt
asgiref==3.5.0
Django==4.0.3
sqlparse==0.4.2
tzdata==2022.1
asgiref==3.5.0
Django==4.0.3
django-ckeditor==6.4.0
django-js-asset==2.0.0
Pillow==9.1.0
sqlparse==0.4.2
tzdata==2022.1
asgiref==3.5.0
Django==4.0.3
django-ckeditor==6.4.0
django-js-asset==2.0.0
Pillow==9.1.0
sqlparse==0.4.2
tzdata==2022.1
